+++ projects/spectator/schematics/src/spectator/schema.ts
@@ -1,4 +1,4 @@
-import { ChangeDetection, ViewEncapsulation } from '@angular/cli/lib/config/schema';
+import { ChangeDetection, ViewEncapsulation } from '@angular/cli/lib/config/workspace-schema';
 
 export class ComponentOptions {
   name: string;
@@ -46,17 +46,9 @@ export class ComponentOptions {
    */
   skipTests?: boolean;
   /**
-   * Specifies if a spec file is generated.
-   */
-  spec?: boolean;
-  /**
    * The file extension or preprocessor to use for style files.
    */
   style?: any;
-  /**
-   * The file extension to be used for style files.
-   */
-  styleext?: string;
   /**
    * Specifies the view encapsulation strategy.
    */
@@ -80,10 +72,6 @@ export class ServiceOptions {
   skipTests?: boolean;
   /**
    * Specifies if a spec file is generated.
-   */
-  spec?: boolean;
-  /**
-   * Specifies if Jest is to be used for mocking
    */
   jest?: boolean;
 }
@@ -119,10 +107,6 @@ export class DirectiveOptions {
    */
   skipTests?: boolean;
   /**
-   * Specifies if a spec file is generated.
-   */
-  spec?: boolean;
-  /**
    * Specifies if Jest is to be used for mocking
    */
   jest?: boolean;
@@ -151,11 +135,12 @@ export class PipeOptions {
    */
   skipTests?: boolean;
   /**
-   * Specifies if a spec file is generated.
-   */
-  spec?: boolean;
-  /**
    * Specifies if Jest is to be used for mocking
    */
   jest?: boolean;
+
+  /**
+   * Adds a developer-defined type to the filename, in the format "name.type.ts"
+   */
+  type?: string;
 }